+++ src/util/reservoir.h
@@ -10,7 +10,7 @@ struct Sample {
     Vector3D x_v, n_v; // visible point and normal (normalized)
     Vector3D x_s, n_s; // sample point and normal (normalized)
     Vector3D L; // outgoing radiance at x_s
-    double pdf; // pdf of the sample
+    float pdf; // pdf of the sample
     Vector3D fcos; // product of bsdf and cosine factor
     Vector3D emittance; // zero + one bounce radiance
     bool is_delta;    // ← new
@@ -39,17 +39,17 @@ DEVICE float inline p_hat(const Sample& s) {
 class Reservoir {
     public:
         Sample z;
-        double w; // sum of weight of samples so far
-        double M; // number of samples so far
-        double W;
+        float w; // sum of weight of samples so far
+        float M; // number of samples so far
+        float W;
 
         HOST_DEVICE Reservoir() : w(0), M(0), W(0) {
             z = Sample();
         }
         
-        DEVICE void update(Sample s_new, double w_new, RNGState &rand_state);
+        DEVICE void update(Sample s_new, float w_new, RNGState &rand_state);
 
-        DEVICE void merge(Reservoir r, double p_hat, RNGState &rand_state);
+        DEVICE void merge(Reservoir r, float p_hat, RNGState &rand_state);
 
         DEVICE void clear();
 };
